requirements.txt
discord.py>=2.5.2
python-dotenv>=1.1.0
feedparser>=6.0.11
beautifulsoup4>=4.13.4 